+++ i7dw/interpro/__init__.py
@@ -1,0 +1,16 @@
+from typing import Tuple
+
+
+DC_STATUSES = {
+    # Continuous single chain domain
+    "S": "CONTINUOUS",
+    # N terminus discontinuous
+    "N": "N_TERMINAL_DISC",
+    # C terminus discontinuous
+    "C": "C_TERMINAL_DISC",
+    # N and C terminus discontinuous
+    "NC": "NC_TERMINAL_DISC"
+}
+
+def repr_frag(f: dict) -> Tuple[int, int]:
+    return f["start"], f["end"]
